POSIX::mblen() Make thread-safe; allow shift state control
This commit changes to use mbrlen() on threaded perls, (mostly)
transparently, when available, to achieve thread-safe operation.  It is
not completely transparent because mbrlen may set errno differently upon
errors, and no effort is made to hide that difference.  Also mbrlen on
some platforms can handle valid, but-partial characters; and the return
value in an edge case may differ.

This commit also changes the behavior so that it takes a scalar
parameter instead of a char *, and this might not be forceable into a
valid PV.  If not a PV, mblen() is always used (with a semaphore) to
initialize the shift state.  Previously the shift state was always
reinitialized with every call, which meant this could not work on
locales with shift states.

[perl #133928] showed that someone was having trouble with shift states.