Collapse grok_bin, _oct, _hex into one function
These functions are identical in logic in the main loop, the difference
being which digits they accept.  The rest of the code had slight
variations.  This commit unifies the functions.

I presume the reason they were kept separate was because of speed.  I
believe I have solved that sufficiently to warrant unifying them, and
thus lowering the maintenance costs.

The key to making this feasible, in my opinion, was changing to compute
the value of a hex digit without branching, done in a previous commit.
That speeds up the grok_hex routine.  The others already didn't have
branching; this commit adds to them a couple of shifts, masks, and
additions per loop iteration, which shouldn't affect the instruction
pipeline.  More importantly, though, it adds an array lookup per
loop iteration to those two routines, which already existed in the
grok_hex one.  I think these two functions are used much less often than
grok_hex, and this minor amount of extra work is justified by the
maintenance cost improvement.  (A branch has been removed from the
grok_bin function loop in compensation, though I suspect many optimizing
compilers would have already removed it.)

This commit does add a few extra conditionals in set-up and on unlikely
error legs.  These should not be noticeable.

The only other thing I can think of that could affect the speed is that
in places a variable is used instead of a constant it replaces.  There
is one instance of this in the loop, involving a shift count.